integration(ai bot) merged ai bot branch to develop
