+++ client/src/App.jsx
@@ -29,4 +29,3 @@ function App() {
 }
 
 export default App;
-export default App;
